[fix]: Code review issues
