+++ packages/base/src/hooks/useProjectTips/index.test.ts
@@ -1,5 +1,5 @@
 import { act, renderHook, cleanup } from '@testing-library/react';
-import useProjects from '.';
+import useProjectTips from '.';
 import project from '../../testUtils/mockApi/project';
 import { mockProjectList } from '../../testUtils/mockApi/project/data';
 import {
@@ -7,7 +7,7 @@ import {
   createSpyFailResponse
 } from '@actiontech/shared/lib/testUtil/mockApi';
 
-describe('useProjects', () => {
+describe('useProjectTips', () => {
   let getProjectListSpy: jest.SpyInstance;
   beforeEach(() => {
     getProjectListSpy = project.getProjectList();
@@ -20,7 +20,7 @@ describe('useProjects', () => {
   });
 
   it('should get projects data from request', async () => {
-    const { result } = renderHook(() => useProjects());
+    const { result } = renderHook(() => useProjectTips());
     expect(result.current.loading).toBe(false);
     expect(result.current.projectList).toEqual([]);
 
@@ -46,7 +46,7 @@ describe('useProjects', () => {
     getProjectListSpy.mockImplementation(() =>
       createSpyFailResponse({ total: 0, users: [] })
     );
-    const { result } = renderHook(() => useProjects());
+    const { result } = renderHook(() => useProjectTips());
     act(() => {
       result.current.updateProjects();
     });
@@ -60,7 +60,7 @@ describe('useProjects', () => {
     getProjectListSpy.mockImplementation(() =>
       createSpyErrorResponse({ total: 0, users: [] })
     );
-    const { result } = renderHook(() => useProjects());
+    const { result } = renderHook(() => useProjectTips());
     act(() => {
       result.current.updateProjects();
     });
