+++ packages/base/src/page/Project/BatchImportDataSource/hooks/useBatchImportDataSource.ts
@@ -38,6 +38,7 @@ const useBatchImportDataSource = () => {
 
   const resetAndHideResult = () => {
     form.resetFields();
+    clearUploadCheckStatus();
     hideResult();
     setDBservices([]);
   };
